refactor(controller): standardize success responses returned across all controllers for all methods
- For ease of dev and consuming api
- Update edit message http method from PUT to PATCH (to better reflect the scope of update)

+++ controller/chats/message/delete.js
@@ -7,6 +7,8 @@ export const deleteMsg = async (req, res, next) => {
   try {
     const msgToDelete = await getMessageById(messageId);
 
+    if (!msgToDelete) return res.status(404).end();
+
     if (msgToDelete.authorId === userId) {
       await deleteMessageById(messageId);
       return res.status(204).end();
